+++ fronts/properties/run.py
@@ -16,13 +16,15 @@ from fronts.properties import algorithms as prop_algorithms
 
 
 def colocate_fronts(timestamp: str, config: str, version: str,
-                    property_names: list, property_dir: str,
+                    property_names: list,
                     output_dir: str = None,
                     stats: list = None, percentiles: list = None,
                     min_npix: int = 1, nan_policy: str = 'omit',
-                    dilation_radius: int = 1, clobber: bool = False,
-                    binary_path: str = None, group_dir: str = None):
+                    dilation_radius: int = 1, clobber: bool = False):
     """Co-locate labeled fronts with physical property fields.
+
+    All paths are resolved from ``PATH/V{version}/YYYYMMDD_HHMMSS/``
+    via :func:`fronts.llc.io.set_fronts_path`.
 
     Args:
         timestamp (str): Snapshot timestamp, e.g. '2012-11-09T12_00_00'.
@@ -32,9 +34,8 @@ def colocate_fronts(timestamp: str, config: str, version: str,
             ['relative_vorticity', 'strain_n']. Each must match both the
             variable name inside its .nc file and the filename pattern
             LLC4320_{timestamp}_{property_name}_{version}.nc.
-        property_dir (str): Directory containing property .nc files.
         output_dir (str, optional): Output directory. Defaults to the
-            standard group_fronts output directory for this version.
+            standard fronts directory for this version + timestamp.
         stats (list, optional): Statistics to compute per property.
             Defaults to ['mean', 'std', 'median'].
         percentiles (list, optional): Percentiles to compute, e.g. [10, 90].
@@ -43,16 +44,12 @@ def colocate_fronts(timestamp: str, config: str, version: str,
         dilation_radius (int): Pixels to dilate each front before stats.
             Defaults to 0.
         clobber (bool): Overwrite existing output. Defaults to False.
-        binary_path (str, optional): Directory for binary front .npy files.
-        group_dir (str, optional): Directory containing group_fronts outputs
-            (label maps). Defaults to ``$OS_OGCM/LLC/Fronts/group_fronts/v{version}``.
     """
-    fronts_file = finding_io.binary_filename(timestamp, config, version, path=binary_path)
-    if group_dir is None:
-        group_dir = os.path.join(os.getenv('OS_OGCM'), 'LLC', 'Fronts',
-                                 'group_fronts', f'v{version}')
+    fdir = llc_io.fronts_dir(version, timestamp)
+    fronts_file = finding_io.binary_filename(timestamp, config, version)
+    property_dir = fdir
     if output_dir is None:
-        output_dir = group_dir
+        output_dir = fdir
 
     # Check if output already exists
     time_str = timestamp.replace('_', ':')   # '2012-11-09T12:00:00'
@@ -78,7 +75,7 @@ def colocate_fronts(timestamp: str, config: str, version: str,
 
     # Load label map
     labeled_file = properties_io.get_global_front_output_path(
-        group_dir, time_str, 'label_map',run_tag)
+        fdir, time_str, 'label_map', run_tag)
     labeled = np.load(labeled_file)
 
     prop_algorithms.colocate_fronts(
@@ -97,14 +94,16 @@ def colocate_fronts(timestamp: str, config: str, version: str,
 
 def generate_properties(timestamp: str, config_file: str, version: str,
                         property_names: list, run_id: str = None,
-                        clobber: bool = False, create_zarr: bool = False,
-                        derived_path: str = None):
+                        clobber: bool = False, create_zarr: bool = False):
     """Generate individual per-property .nc files for the requested properties.
 
     Resolves which dbof subset each property belongs to from the YAML config,
     then writes one LLC4320_{timestamp}_{property}_v{version}.nc file per
     property — the format expected by colocate_fronts(). Existing files are
     skipped unless clobber=True.
+
+    Use :func:`fronts.llc.io.set_fronts_path` to override the root
+    directory.  Files land under ``PATH/V{version}/YYYYMMDD_HHMMSS/``.
 
     Args:
         timestamp (str): Snapshot timestamp, e.g. '2012-11-09T12_00_00'.
@@ -116,8 +115,6 @@ def generate_properties(timestamp: str, config_file: str, version: str,
         clobber (bool): Overwrite existing output files. Defaults to False.
         create_zarr (bool): Create the zarr store via generate_global.
             Defaults to False (assumes zarr already exists on S3).
-        derived_path (str, optional): Override the output directory for
-            derived .nc files.  Defaults to ``$OS_OGCM/LLC/Fronts/derived``.
     """
     with open(config_file) as fh:
         raw = yaml.safe_load(fh) or {}
@@ -151,7 +148,7 @@ def generate_properties(timestamp: str, config_file: str, version: str,
     # Process each subset
     for subset, channels in subset_to_channels.items():
         missing = [ch for ch in channels
-                   if not os.path.isfile(llc_io.derived_filename(timestamp, ch, version=version, path=derived_path))]
+                   if not os.path.isfile(llc_io.derived_filename(timestamp, ch, version=version))]
 
         if not missing and not clobber:
             print(f"All {len(channels)} property file(s) for subset '{subset}' exist "
@@ -168,12 +165,14 @@ def generate_properties(timestamp: str, config_file: str, version: str,
         # Convert zarr → netcdf for each channel
         for channel in to_generate:
             llc_io.zarr_to_nc(timestamp, config_file, subset, field=channel,
-                        version=version, run_id=run_id, path=derived_path)
+                        version=version, run_id=run_id)
 
 def group_fronts(timestamp: str, config: str, version: str,
-                 n_workers: int = None, skip_curvature: bool = False,
-                 binary_path: str = None, output_dir: str = None):
+                 n_workers: int = None, skip_curvature: bool = False):
     """Label connected front components and compute geometric properties globally.
+
+    All paths are resolved from ``PATH/V{version}/YYYYMMDD_HHMMSS/``
+    via :func:`fronts.llc.io.set_fronts_path`.
 
     Args:
         timestamp (str): Snapshot timestamp, e.g. '2012-11-09T12_00_00'.
@@ -181,14 +180,10 @@ def group_fronts(timestamp: str, config: str, version: str,
         version (str): Data version string.
         n_workers (int, optional): Parallel workers. Defaults to CPU count.
         skip_curvature (bool): Skip curvature calculation (~50% faster).
-        binary_path (str, optional): Directory for binary front .npy files.
-        output_dir (str, optional): Output directory for group_fronts results.
     """
-    fronts_file = finding_io.binary_filename(timestamp, config, version, path=binary_path)
+    fronts_file = finding_io.binary_filename(timestamp, config, version)
     coords_file = os.path.join(os.getenv('OS_OGCM'), 'LLC', 'Fronts', 'coords', 'LLC_coords_lat_lon.nc')
-    if output_dir is None:
-        output_dir = os.path.join(os.getenv('OS_OGCM'), 'LLC', 'Fronts',
-                                  'group_fronts', f'v{version}')
+    output_dir = llc_io.fronts_dir(version, timestamp)
 
     # Load
     fronts_binary = np.load(fronts_file)
