+++ fronts/runs/prototypes/one_full/build_v3.py
@@ -4,6 +4,8 @@
 import os
 import sys
 
+from fronts.llc import io as llc_io
+
 from fronts.preproc.gradb2 import generate_gradb2
 from fronts.finding.run import find_gradb2_fronts
 from fronts.properties.run import group_fronts
@@ -23,29 +25,22 @@ def main(flg: str):
     config_file = './testing_global_v3.yaml'
     run_id      = 'global_20121109_120000_v3'
 
-    # Versioned directory layout:
-    #   $OS_OGCM/LLC/Fronts/v3/derived/
-    #   $OS_OGCM/LLC/Fronts/v3/outputs/
-    #   $OS_OGCM/LLC/Fronts/v3/group_fronts/
-    fronts_root  = os.path.join(os.getenv('OS_OGCM'), 'LLC', 'Fronts', f'v{version}')
-    derived_dir  = os.path.join(fronts_root, 'derived')
-    outputs_dir  = os.path.join(fronts_root, 'outputs')
-    group_dir    = os.path.join(fronts_root, 'group_fronts')
+    # All products land under: PATH / V{version} / YYYYMMDD_HHMMSS /
+    # e.g. $OS_OGCM/LLC/Fronts/V3/20121109_120000/
+    llc_io.set_fronts_path(os.path.join(os.getenv('OS_OGCM'), 'LLC', 'Fronts'))
 
     # Generate gradb2 as zarr → netcdf
     if flg == 1:
         generate_gradb2(timestamp, config_file, version=version, run_id=run_id,
-            create_zarr=False, path=derived_dir)
+            create_zarr=False)
 
     # Find fronts -- binary pixels
     if flg == 2:
-        find_gradb2_fronts(timestamp, config, version,
-            derived_path=derived_dir, output_path=outputs_dir)
+        find_gradb2_fronts(timestamp, config, version)
 
     # Group fronts (label + geometric properties)
     if flg == 3:
-        group_fronts(timestamp, config, version,
-            binary_path=outputs_dir, output_dir=group_dir)
+        group_fronts(timestamp, config, version)
 
     # Generate physical property fields → individual per-property nc files
     if flg == 4:
@@ -60,7 +55,7 @@ def main(flg: str):
         ]
         generate_properties(timestamp, config_file, version,
                             property_names=property_names, run_id=run_id,
-                            create_zarr=False, derived_path=derived_dir)
+                            create_zarr=False)
 
     # Co-locate fronts with physical properties
     if flg == 5:
@@ -74,10 +69,7 @@ def main(flg: str):
             'oceTAUX','oceTAUY','SIarea',
         ]
         colocate_fronts(timestamp, config, version,
-                            property_names=property_names,
-                            property_dir=derived_dir,
-                            binary_path=outputs_dir,
-                            group_dir=group_dir)
+                            property_names=property_names)
 
 
 # Command line execution
